Fix all remaining compilation errors
Fixed Issues:
1. Packet undeclared - Changed to NetworkClient::ReceivedPacket
2. receivePacket() - Changed to hasPackets()/getNextPacket()
3. strncpy_s arguments - Added buffer size parameter (3 args not 2)
4. password field - Changed to passwordHash in LoginRequest/RegisterRequest
5. message field - Changed to errorMessage in LoginResponse/RegisterResponse
6. packet.data - Changed to packet.payload.data()
7. Double to float math - Changed std::sin/cos/sqrt/pow to sinf/cosf/sqrtf/powf
8. M_PI constant - Added 'f' suffix for float literal
9. uint64_t to uint32_t - Added static_cast for all sizeof() in server

Changes by file:
- LoginScene.cpp: Fixed packet handling, strncpy_s, field names
- LoginUI.cpp: Fixed strncpy_s calls with buffer size
- Animation.cpp: Replaced all double math with float versions
- MainMenuScene.cpp, MainMenuUI.cpp, UIButton.cpp: Changed std::sin to std::sinf
- server/main.cpp: Added static_cast<uint32_t> to all sizeof()

All networking code now uses correct types and function signatures.
All math operations use float versions to avoid double→float warnings.

+++ src/client/ui/MainMenuScene.cpp
@@ -198,7 +198,7 @@ void MainMenuScene::update(float deltaTime) {
     animTime += deltaTime;
 
     // Animate title
-    float pulse = 0.85f + 0.15f * std::sin(animTime * 2.0f);
+    float pulse = 0.85f + 0.15f * std::sinf(animTime * 2.0f);
     titleText->setColor(Color(pulse, pulse * 0.85f, pulse * 0.5f, 1.0f));
 
     // Update info text based on hovered button
